% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optsoln.R
\name{schedule}
\alias{schedule}
\title{Get linearly scaled variable based on optimal bins}
\usage{
schedule(data)
}
\description{
Finds outliers and then bands between 0 and 1 on optimal bins of
non-outlier data
}
